+++ src/protocols/kubernetes/clipboard.c
@@ -20,6 +20,7 @@
 #include "clipboard.h"
 #include "common/clipboard.h"
 #include "kubernetes.h"
+#include "terminal/terminal_priv.h"
 
 #include <guacamole/client.h>
 #include <guacamole/stream.h>
@@ -33,7 +34,7 @@ int guac_kubernetes_clipboard_handler(guac_user* user, guac_stream* stream,
         (guac_kubernetes_client*) client->data;
 
     /* Clear clipboard and prepare for new data */
-    guac_common_clipboard_reset(kubernetes_client->clipboard, mimetype);
+    guac_common_clipboard_reset(kubernetes_client->term->clipboard, mimetype);
 
     /* Set handlers for clipboard stream */
     stream->blob_handler = guac_kubernetes_clipboard_blob_handler;
@@ -50,7 +51,7 @@ int guac_kubernetes_clipboard_blob_handler(guac_user* user,
         (guac_kubernetes_client*) client->data;
 
     /* Append new data */
-    guac_common_clipboard_append(kubernetes_client->clipboard, data, length);
+    guac_common_clipboard_append(kubernetes_client->term->clipboard, data, length);
 
     return 0;
 }
