+++ src/protocols/kubernetes/kubernetes.c
@@ -26,6 +26,7 @@
 #include "kubernetes.h"
 #include "ssl.h"
 #include "terminal/terminal.h"
+#include "terminal/terminal_priv.h"
 #include "url.h"
 
 #include <guacamole/client.h>
@@ -241,8 +242,7 @@ void* guac_kubernetes_client_thread(void* data) {
 
     /* Create terminal options with required parameters */
     guac_terminal_options* options = guac_terminal_options_create(
-            client, kubernetes_client->clipboard,
-            settings->width, settings->height, settings->resolution);
+            client, settings->width, settings->height, settings->resolution);
 
     /* Set optional parameters */
     options->disable_copy = settings->disable_copy;
