+++ src/protocols/rdp/client.h
@@ -63,11 +63,6 @@
 #define GUAC_RDP_REASONABLE_AREA (800*600)
 
 /**
- * The maximum number of bytes to allow within the clipboard.
- */
-#define GUAC_RDP_CLIPBOARD_MAX_LENGTH 262144
-
-/**
  * Initial rate of audio to stream, in Hz. If the RDP server uses a different
  * value, the Guacamole audio stream will simply be reset appropriately.
  */
