+++ src/protocols/ssh/client.c
@@ -21,7 +21,6 @@
 
 #include "argv.h"
 #include "client.h"
-#include "common/clipboard.h"
 #include "common/recording.h"
 #include "common-ssh/sftp.h"
 #include "ssh.h"
@@ -44,9 +43,6 @@ int guac_client_init(guac_client* client) {
     /* Allocate client instance data */
     guac_ssh_client* ssh_client = calloc(1, sizeof(guac_ssh_client));
     client->data = ssh_client;
-
-    /* Init clipboard */
-    ssh_client->clipboard = guac_common_clipboard_alloc(GUAC_SSH_CLIPBOARD_MAX_LENGTH);
 
     /* Set handlers */
     client->join_handler = guac_ssh_user_join_handler;
@@ -118,7 +114,6 @@ int guac_ssh_client_free_handler(guac_client* client) {
         guac_ssh_settings_free(ssh_client->settings);
 
     /* Free client structure */
-    guac_common_clipboard_free(ssh_client->clipboard);
     free(ssh_client);
 
     guac_common_ssh_uninit();
