+++ src/protocols/ssh/ssh.c
@@ -27,6 +27,7 @@
 #include "sftp.h"
 #include "ssh.h"
 #include "terminal/terminal.h"
+#include "terminal/terminal_priv.h"
 #include "ttymode.h"
 
 #ifdef ENABLE_SSH_AGENT
@@ -240,8 +241,7 @@ void* ssh_client_thread(void* data) {
 
     /* Create terminal options with required parameters */
     guac_terminal_options* options = guac_terminal_options_create(
-            client, ssh_client->clipboard,
-            settings->width, settings->height, settings->resolution);
+            client, settings->width, settings->height, settings->resolution);
 
     /* Set optional parameters */
     options->disable_copy = settings->disable_copy;
