+++ src/protocols/telnet/input.c
@@ -21,6 +21,7 @@
 #include "common/recording.h"
 #include "input.h"
 #include "terminal/terminal.h"
+#include "terminal/terminal_priv.h"
 #include "telnet.h"
 
 #include <guacamole/client.h>
