+++ src/protocols/telnet/user.c
@@ -26,6 +26,7 @@
 #include "settings.h"
 #include "telnet.h"
 #include "terminal/terminal.h"
+#include "terminal/terminal_priv.h"
 #include "user.h"
 
 #include <guacamole/client.h>
