+++ src/terminal/terminal-stdin-stream.c
@@ -20,6 +20,7 @@
 #include "config.h"
 #include "terminal/common.h"
 #include "terminal/terminal.h"
+#include "terminal/terminal_priv.h"
 
 #include <guacamole/protocol.h>
 #include <guacamole/socket.h>
