+++ src/terminal/terminal_handlers.c
@@ -23,6 +23,7 @@
 #include "terminal/palette.h"
 #include "terminal/terminal.h"
 #include "terminal/terminal_handlers.h"
+#include "terminal/terminal_priv.h"
 #include "terminal/types.h"
 #include "terminal/xparsecolor.h"
